direct access to semaphore without decorator prohibited
